+++ packages/harch-web/test/Unit/HarchWeb/SiteSpec.hs
@@ -468,8 +468,9 @@ waiRequest segments =
         [] -> "/"
         _ -> "/" <> Text.intercalate "/" segments
 
-performWaiRequest :: Wai.Application -> Wai.Request -> IO Wai.Response
-performWaiRequest webApplication request = do
+performWaiRequest :: IO Wai.Application -> Wai.Request -> IO Wai.Response
+performWaiRequest buildWebApplication request = do
+  webApplication <- buildWebApplication
   responseReference <- newIORef Nothing
   _ <-
     webApplication