+++ packages/web-api/test/Unit/WebApiSpec.hs
@@ -403,8 +403,9 @@ data CapturedOtlpRequest = CapturedOtlpRequest
     capturedOtlpBody :: ByteString.ByteString
   }
 
-performWaiRequest :: Wai.Application -> Wai.Request -> IO Wai.Response
-performWaiRequest webApplication request = do
+performWaiRequest :: IO Wai.Application -> Wai.Request -> IO Wai.Response
+performWaiRequest buildWebApplication request = do
+  webApplication <- buildWebApplication
   responseReference <- newIORef Nothing
   _ <- webApplication request (\response -> writeIORef responseReference (Just response) >> pure WaiInternal.ResponseReceived)
   maybeResponse <- readIORef responseReference